requirements.txt
Flask~=3.0.1
flask-paginate~=0.6.0
requests~=2.25.0
beautifulsoup4~=4.12.3
bs4~=0.0.2
lxml~=5.1.0
soupsieve~=2.5
Werkzeug~=3.0.1
Jinja2~=3.1.3
itsdangerous~=2.1.2
MarkupSafe~=2.1.4
PyYAML~=6.0.1
blinker~=1.7.0
click~=8.1.7
pygments~=2.17.2
pyserial~=3.5
